Split InstanceSerializationTest and move tests to corresponding feature files
