+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/formmanagement/BulkFinalizationTest.kt
@@ -4,6 +4,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.equalTo
 import org.hamcrest.Matchers.not
+import org.javarosa.xform.parse.XFormParser
 import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.RuleChain
@@ -280,5 +281,30 @@ class BulkFinalizationTest {
         assertThat(firstFormDeprecatedID, equalTo(null))
         assertThat(firstFormInstanceID, equalTo(secondFormDeprecatedID))
         assertThat(secondFormInstanceID, not(firstFormInstanceID))
+    }
+
+    @Test
+    fun finalizingAllDrafts_prunesNonRelevantNodes() {
+        testDependencies.server.addForm("one-question-relevance.xml")
+
+        rule.withProject(testDependencies.server.url, matchExactly = true)
+            .startBlankForm("One Question Relevance")
+            .clickOnText("Yes")
+            .swipeToNextQuestion("what is your age")
+            .swipeToPreviousQuestion("Do you want to continue?")
+            .clickOnText("No")
+            .swipeToEndScreen()
+            .clickSaveAsDraft()
+            .clickDrafts(1)
+            .clickFinalizeAll(1)
+            .clickFinalize()
+            .pressBack(MainMenuPage())
+            .clickSendFinalizedForm(1)
+            .clickSelectAll()
+            .clickSendSelected()
+
+        val file = testDependencies.server.submissions[0]
+        val instanceRootElement = XFormParser.getXMLDocument(file.inputStream().reader()).rootElement
+        assertThat(instanceRootElement.indexOf(null, "age", 0), equalTo(-1))
     }
 }
