+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/instancemanagement/SendFinalizedFormTest.kt
@@ -184,4 +184,25 @@ class SendFinalizedFormTest {
         assertThat((firstFormRootElement.getChild(0) as Element).getChild(0), equalTo("123"))
         assertThat((secondFormRootElement.getChild(0) as Element).getChild(0), equalTo("124"))
     }
+
+    @Test
+    fun sentFormDoesNotIncludeNonRelevantNodes() {
+        testDependencies.server.addForm("one-question-relevance.xml")
+
+        rule.withProject(testDependencies.server.url, matchExactly = true)
+            .startBlankForm("One Question Relevance")
+            .clickOnText("Yes")
+            .swipeToNextQuestion("what is your age")
+            .swipeToPreviousQuestion("Do you want to continue?")
+            .clickOnText("No")
+            .swipeToEndScreen()
+            .clickFinalize()
+            .clickSendFinalizedForm(1)
+            .clickSelectAll()
+            .clickSendSelected()
+
+        val file = testDependencies.server.submissions[0]
+        val instanceRootElement = XFormParser.getXMLDocument(file.inputStream().reader()).rootElement
+        assertThat(instanceRootElement.indexOf(null, "age", 0), equalTo(-1))
+    }
 }
